# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Patient Tracker App",
        "contact": {
            "name": "Mahmoud Mohamed",
            "url": "https://www.linkedin.com/in/mahmoud-mohamed-955932b5/",
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0"
    },
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/patients": {
            "get": {
                "tags": [
                    "patients"
                ],
                "summary": "get all patients paginated",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "no update , confirmed , not ph",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "city_id",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "country_id",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "hospital_id",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "doctor_id",
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "hospital_type",
                        "in": "query",
                        "description": "coe , referal",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "add new patient",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "city_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "country_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "hospital_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "doctor_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type_medication",
                        "in": "formData",
                        "description": "uptravi , opsumit",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "etiology",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "other_medication",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/patients/{patient}": {
            "get": {
                "tags": [
                    "patients"
                ],
                "summary": "get single patient",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/patients/{patient}/treatments": {
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "add new treatments",
                "parameters": [
                    {
                        "name": "patient",
                        "in": "path",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "type_medication",
                        "in": "formData",
                        "description": "uptravi , opsumit",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "etiology",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "other_medication",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/patients/add-referal": {
            "post": {
                "tags": [
                    "patients"
                ],
                "summary": "add new refaral patient",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "city_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "country_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "hospital_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "doctor_id",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "to_hospital",
                        "in": "formData",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/me": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Get the current logged in user",
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Account",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoud"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "rep1@gmail.com"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update-password": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Password",
                "parameters": [
                    {
                        "name": "current_password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Model"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}